Struct tower_util::boxed::BoxService

pub struct BoxService<T, U, E> { /* fields omitted */ }

A boxed Service + Send trait object.

BoxService turns a service into a trait object, allowing the response future type to be dynamic. This type requires both the service and the response future to be Send.

See module level documentation for more details.

Methods

impl<T, U, E> BoxService<T, U, E>

pub fn new<S>(inner: S) -> Self where
S: Service<Request = T, Response = U, Error = E> + Send + 'static,
S::Future: Send + 'static,