Enum tower_util::either::EitherService

pub enum EitherService<A, B> {
    A(A),
    B(B),
}

Combine two different service types into a single type.

Both services must be of the same request, response, and error types. EitherService is useful for handling conditional branching in service middleware to different inner service types.

Variants
A(A)B(B)