Trait tower_util::MakeConnection

pub trait MakeConnection<Request> {
    type Response: AsyncRead + AsyncWrite;
    type Error;
    type Future: Future<Item = Self::Response, Error = Self::Error>;
    fn make_connection(&mut self, target: Request) -> Self::Future;
}

The MakeConnection trait is used to create transports

The goal of this service is to allow composable methods for creating AsyncRead + AsyncWrite transports. This could mean creating a TLS based connection or using some other method to authenticate the connection.

Associated Types

type Response: AsyncRead + AsyncWrite

The transport provided by this service

type Error

Errors produced by the connecting service

type Future: Future<Item = Self::Response, Error = Self::Error>

The future that eventually produces the transport

Required Methods

fn make_connection(&mut self, target: Request) -> Self::Future

Connect and return a transport asynchronously