Trait tower_util::ServiceExt

pub trait ServiceExt<Request>: Service<Request> {
    fn ready(&mut self) -> Ready<Self, Request>
    where
        Self: Sized,
    { ... }

    fn ready_and(&mut self) -> ReadyAnd<Self, Request>
    where
        Self: Sized,
    { ... }

    fn ready_oneshot(self) -> ReadyOneshot<Self, Request>
    where
        Self: Sized,
    { ... }

    fn oneshot(self, req: Request) -> Oneshot<Self, Request>
    where
        Self: Sized,
    { ... }

    fn call_all<S>(self, reqs: S) -> CallAll<Self, S>
    where
        Self: Sized,
        Self::Error: Into<Error>,
        S: Stream<Item = Request>,
    { ... }
}

An extension trait for Services that provides a variety of convenient adapters

Provided methods

fn ready(&mut self) -> Ready<Self, Request> where
Self: Sized,

Deprecated since 0.3.1:
prefer ready_and which yields the service

Resolves when the service is ready to accept a request.

fn ready_and(&mut self) -> ReadyAnd<Self, Request> where
Self: Sized,

Yields a mutable reference to the service when it is ready to accept a request.

fn ready_oneshot(self) -> ReadyOneshot<Self, Request> where
Self: Sized,

Yields the service when it is ready to accept a request.

fn oneshot(self, req: Request) -> Oneshot<Self, Request> where
Self: Sized,

Consume this Service, calling with the providing request once it is ready.

fn call_all<S>(self, reqs: S) -> CallAll<Self, S> where
Self: Sized,
Self::Error: Into<Error>,
S: Stream<Item = Request>,

Process all requests from the given Stream, and produce a Stream of their responses.

This is essentially Stream<Item = Request> + Self => Stream<Item = Response>. See the documentation for CallAll for details.

Loading content...